Massive Overhaul follow-up
-> Trashed completely previous Scale/Mirror/Flips function from sprites.
Manipulating pixels at runtime for simple operations like these can only
lead to disaster.

-> Implemented native sprite scaling/flipping/rotation. Works with
animations. Documentation has been added inside Sprite.cpp, it should be
quite easy to use.

-> Cleaned up impertinent functions, variables and comments across
several files.

-> Standardized function arguments names & order across different files.

Warning:
-> Pixel Manipulation remains unavailable for the time being. Needs a
valid reason to consider re-implementation.

More documentation and examples should be available soon.

+++ src/Core/Text.cpp
@@ -208,7 +208,7 @@ void Text::UpdateFadeIn()
 	{
 		if (fadeInTxtsList[i].used)
 		{
-			fadeInTxtsList[i].currentTime += dt;
+			fadeInTxtsList[i].currentTime += (int)dt;
 			if (fadeInTxtsList[i].currentTime >= 1.0f / framerate)
 			{
 				fadeInTxtsList[i].currentTime = 0;
